fix: msgSeq and msgRandom type (#62)
* fix: msgSeq and msgRandom type

* build: release v0.2.8

* docs: add code comments

+++ src/main/java/io/github/doocs/im/model/group/MemberProfile.java
@@ -20,8 +20,11 @@ public class MemberProfile {
     @JsonProperty("JoinTime")
     private Integer joinTime;
 
+    /**
+     * 消息序列号，用于标记该条消息，数据范围 [0,4294967295]
+     */
     @JsonProperty("MsgSeq")
-    private Integer msgSeq;
+    private Long msgSeq;
 
     @JsonProperty("MsgFlag")
     private String msgFlag;
@@ -41,7 +44,7 @@ public class MemberProfile {
     public MemberProfile() {
     }
 
-    public MemberProfile(String memberAccount, String role, Integer joinTime, Integer msgSeq,
+    public MemberProfile(String memberAccount, String role, Integer joinTime, Long msgSeq,
                          String msgFlag, Integer lastSendMsgTime, Integer shutUpUntil, String nameCard,
                          List<AppMemberDefinedDataItem> appMemberDefinedData) {
         this.memberAccount = memberAccount;
@@ -95,11 +98,11 @@ public class MemberProfile {
         this.joinTime = joinTime;
     }
 
-    public Integer getMsgSeq() {
+    public Long getMsgSeq() {
         return msgSeq;
     }
 
-    public void setMsgSeq(Integer msgSeq) {
+    public void setMsgSeq(Long msgSeq) {
         this.msgSeq = msgSeq;
     }
 
@@ -162,7 +165,10 @@ public class MemberProfile {
         private String memberAccount;
         private String role;
         private Integer joinTime;
-        private Integer msgSeq;
+        /**
+         * 消息序列号，用于标记该条消息，数据范围 [0,4294967295]
+         */
+        private Long msgSeq;
         private String msgFlag;
         private Integer lastSendMsgTime;
         private Integer shutUpUntil;
@@ -191,7 +197,7 @@ public class MemberProfile {
             return this;
         }
 
-        public Builder msgSeq(Integer msgSeq) {
+        public Builder msgSeq(Long msgSeq) {
             this.msgSeq = msgSeq;
             return this;
         }
