+++ src/main/java/io/github/doocs/im/model/request/BatchSendMsgRequest.java
@@ -21,11 +21,17 @@ public class BatchSendMsgRequest {
     @JsonProperty("To_Account")
     private List<String> toAccount;
 
+    /**
+     * 消息序列号，用于标记该条消息，数据范围 [0,4294967295]
+     */
     @JsonProperty("MsgSeq")
-    private Integer msgSeq;
+    private Long msgSeq;
 
+    /**
+     * 消息随机数，用于标记该条消息，数据范围 [0,4294967295]
+     */
     @JsonProperty("MsgRandom")
-    private Integer msgRandom;
+    private Long msgRandom;
 
     @JsonProperty("MsgBody")
     private List<TIMMsgElement> msgBody;
@@ -42,14 +48,14 @@ public class BatchSendMsgRequest {
     public BatchSendMsgRequest() {
     }
 
-    public BatchSendMsgRequest(List<String> toAccount, Integer msgRandom, List<TIMMsgElement> msgBody) {
+    public BatchSendMsgRequest(List<String> toAccount, Long msgRandom, List<TIMMsgElement> msgBody) {
         this.toAccount = toAccount;
         this.msgRandom = msgRandom;
         this.msgBody = msgBody;
     }
 
-    public BatchSendMsgRequest(Integer syncOtherMachine, String fromAccount, List<String> toAccount, Integer msgSeq,
-                               Integer msgRandom, List<TIMMsgElement> msgBody, String cloudCustomData,
+    public BatchSendMsgRequest(Integer syncOtherMachine, String fromAccount, List<String> toAccount, Long msgSeq,
+                               Long msgRandom, List<TIMMsgElement> msgBody, String cloudCustomData,
                                List<String> sendMsgControl, OfflinePushInfo offlinePushInfo) {
         this.syncOtherMachine = syncOtherMachine;
         this.fromAccount = fromAccount;
@@ -102,19 +108,19 @@ public class BatchSendMsgRequest {
         this.toAccount = toAccount;
     }
 
-    public Integer getMsgSeq() {
+    public Long getMsgSeq() {
         return msgSeq;
     }
 
-    public void setMsgSeq(Integer msgSeq) {
+    public void setMsgSeq(Long msgSeq) {
         this.msgSeq = msgSeq;
     }
 
-    public Integer getMsgRandom() {
+    public Long getMsgRandom() {
         return msgRandom;
     }
 
-    public void setMsgRandom(Integer msgRandom) {
+    public void setMsgRandom(Long msgRandom) {
         this.msgRandom = msgRandom;
     }
 
@@ -150,13 +156,18 @@ public class BatchSendMsgRequest {
         this.offlinePushInfo = offlinePushInfo;
     }
 
-
     public static final class Builder {
         private Integer syncOtherMachine;
         private String fromAccount;
         private List<String> toAccount;
-        private Integer msgSeq;
-        private Integer msgRandom;
+        /**
+         * 消息序列号，用于标记该条消息，数据范围 [0,4294967295]
+         */
+        private Long msgSeq;
+        /**
+         * 消息随机数，用于标记该条消息，数据范围 [0,4294967295]
+         */
+        private Long msgRandom;
         private List<TIMMsgElement> msgBody;
         private String cloudCustomData;
         private List<String> sendMsgControl;
@@ -184,12 +195,12 @@ public class BatchSendMsgRequest {
             return this;
         }
 
-        public Builder msgSeq(Integer msgSeq) {
+        public Builder msgSeq(Long msgSeq) {
             this.msgSeq = msgSeq;
             return this;
         }
 
-        public Builder msgRandom(Integer msgRandom) {
+        public Builder msgRandom(Long msgRandom) {
             this.msgRandom = msgRandom;
             return this;
         }
