+++ src/main/java/io/github/doocs/im/model/request/ImportMsgRequest.java
@@ -21,11 +21,17 @@ public class ImportMsgRequest extends GenericRequest {
     @JsonProperty("To_Account")
     private String toAccount;
 
+    /**
+     * 消息序列号，用于标记该条消息，数据范围 [0,4294967295]
+     */
     @JsonProperty("MsgSeq")
-    private Integer msgSeq;
+    private Long msgSeq;
 
+    /**
+     * 消息随机数，用于标记该条消息，数据范围 [0,4294967295]
+     */
     @JsonProperty("MsgRandom")
-    private Integer msgRandom;
+    private Long msgRandom;
 
     @JsonProperty("MsgTimeStamp")
     private Integer msgTimeStamp;
@@ -39,7 +45,7 @@ public class ImportMsgRequest extends GenericRequest {
     public ImportMsgRequest() {
     }
 
-    public ImportMsgRequest(Integer syncFromOldSystem, String fromAccount, String toAccount, Integer msgRandom,
+    public ImportMsgRequest(Integer syncFromOldSystem, String fromAccount, String toAccount, Long msgRandom,
                             Integer msgTimeStamp, List<TIMMsgElement> msgBody) {
         this.syncFromOldSystem = syncFromOldSystem;
         this.fromAccount = fromAccount;
@@ -49,8 +55,8 @@ public class ImportMsgRequest extends GenericRequest {
         this.msgBody = msgBody;
     }
 
-    public ImportMsgRequest(Integer syncFromOldSystem, String fromAccount, String toAccount, Integer msgSeq,
-                            Integer msgRandom, Integer msgTimeStamp, List<TIMMsgElement> msgBody, String cloudCustomData) {
+    public ImportMsgRequest(Integer syncFromOldSystem, String fromAccount, String toAccount, Long msgSeq,
+                            Long msgRandom, Integer msgTimeStamp, List<TIMMsgElement> msgBody, String cloudCustomData) {
         this.syncFromOldSystem = syncFromOldSystem;
         this.fromAccount = fromAccount;
         this.toAccount = toAccount;
@@ -100,19 +106,19 @@ public class ImportMsgRequest extends GenericRequest {
         this.toAccount = toAccount;
     }
 
-    public Integer getMsgSeq() {
+    public Long getMsgSeq() {
         return msgSeq;
     }
 
-    public void setMsgSeq(Integer msgSeq) {
+    public void setMsgSeq(Long msgSeq) {
         this.msgSeq = msgSeq;
     }
 
-    public Integer getMsgRandom() {
+    public Long getMsgRandom() {
         return msgRandom;
     }
 
-    public void setMsgRandom(Integer msgRandom) {
+    public void setMsgRandom(Long msgRandom) {
         this.msgRandom = msgRandom;
     }
 
@@ -140,13 +146,18 @@ public class ImportMsgRequest extends GenericRequest {
         this.cloudCustomData = cloudCustomData;
     }
 
-
     public static final class Builder {
         private Integer syncFromOldSystem;
         private String fromAccount;
         private String toAccount;
-        private Integer msgSeq;
-        private Integer msgRandom;
+        /**
+         * 消息序列号，用于标记该条消息，数据范围 [0,4294967295]
+         */
+        private Long msgSeq;
+        /**
+         * 消息随机数，用于标记该条消息，数据范围 [0,4294967295]
+         */
+        private Long msgRandom;
         private Integer msgTimeStamp;
         private List<TIMMsgElement> msgBody;
         private String cloudCustomData;
@@ -173,12 +184,12 @@ public class ImportMsgRequest extends GenericRequest {
             return this;
         }
 
-        public Builder msgSeq(Integer msgSeq) {
+        public Builder msgSeq(Long msgSeq) {
             this.msgSeq = msgSeq;
             return this;
         }
 
-        public Builder msgRandom(Integer msgRandom) {
+        public Builder msgRandom(Long msgRandom) {
             this.msgRandom = msgRandom;
             return this;
         }
