+++ src/main/java/io/github/doocs/im/model/response/MsgListItem.java
@@ -16,11 +16,17 @@ public class MsgListItem {
     @JsonProperty("To_Account")
     private String toAccount;
 
+    /**
+     * 消息序列号，用于标记该条消息，数据范围 [0,4294967295]
+     */
     @JsonProperty("MsgSeq")
-    private Integer msgSeq;
+    private Long msgSeq;
 
+    /**
+     * 消息随机数，用于标记该条消息，数据范围 [0,4294967295]
+     */
     @JsonProperty("MsgRandom")
-    private Integer msgRandom;
+    private Long msgRandom;
 
     @JsonProperty("MsgTimeStamp")
     private Integer msgTimeStamp;
@@ -53,19 +59,19 @@ public class MsgListItem {
         this.toAccount = toAccount;
     }
 
-    public Integer getMsgSeq() {
+    public Long getMsgSeq() {
         return msgSeq;
     }
 
-    public void setMsgSeq(Integer msgSeq) {
+    public void setMsgSeq(Long msgSeq) {
         this.msgSeq = msgSeq;
     }
 
-    public Integer getMsgRandom() {
+    public Long getMsgRandom() {
         return msgRandom;
     }
 
-    public void setMsgRandom(Integer msgRandom) {
+    public void setMsgRandom(Long msgRandom) {
         this.msgRandom = msgRandom;
     }
